+++ packages/usdk/packages/upstreet-agent/packages/react-agents/index.tsx
@@ -1,16 +1,16 @@
 // components
 
-export * from './components';
-export * from './default-components';
+export * from './components/base-components';
+export * from './components/default-components';
 export * from './hooks';
 
 // types
 
-export type {
-  AgentEvent,
-  PendingActionEvent,
-  PerceptionEvent,
-  ActionMessage,
-  SubtleAiCompleteOpts,
-  AgentAppProps,
-} from './types';
+// export type {
+//   AgentEvent,
+//   PendingActionEvent,
+//   PerceptionEvent,
+//   ActionMessage,
+//   SubtleAiCompleteOpts,
+//   AgentAppProps,
+// } from './types';
