+++ packages/usdk/packages/upstreet-agent/packages/react-agents/loops/chat-loop.tsx
@@ -1,6 +1,6 @@
 import React, { useState } from 'react';
 import { useAgent } from '../hooks';
-import { Perception } from '../components';
+import { Perception } from '../components/base-components';
 import { LoopProps } from './types';
 import { BasicEvaluator } from '../evaluators/basic-evaluator';
